EncryptionServer/requirements.txt
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
des==1.0.5
future==0.17.1
idna==2.8
iso8601==0.1.12
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
Naked==0.1.31
netifaces==0.10.6
numpy==1.16.4
opencv-python==4.1.0.25
passlib==1.7.1
pycryptodome==3.8.2
pylint==2.3.1
PyQt5==5.11.3
PyQt5-sip==4.19.17
pyqt5-tools==5.11.3.1.4
pyserial==3.4
python-dotenv==0.10.3
python-engineio==3.8.1
python-socketio==4.1.0
PyYAML==5.1.1
requests==2.22.0
serial==0.0.97
shellescape==3.4.1
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.3
websocket-client==0.56.0
wrapt==1.11.1
